Created basic player and bullet prefabs, added simple network manager.
So far in the progress of this tutorial we have created simple player representations as capsules with black glasses and a cylinder for a gun. We used a reduced scale sphere object as the bullet prefab, and we have established a simple network manager so that we can play each other locally between two separate builds. The players can move and fire the weapon, but no reactions have been established to getting hit by a bullet.